ProjectManager and input.io moved to common package
