+++ web-ui/src/test/java/org/archcnl/domain/input/io/AdocEndToEndTest.java
@@ -8,6 +8,8 @@ import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import org.apache.commons.io.FileUtils;
 import org.archcnl.domain.TestUtils;
+import org.archcnl.domain.common.io.ArchRulesFromAdocReader;
+import org.archcnl.domain.common.io.ArchRulesToAdocWriter;
 import org.archcnl.domain.input.exceptions.ConceptDoesNotExistException;
 import org.archcnl.domain.input.exceptions.RelationDoesNotExistException;
 import org.archcnl.domain.input.model.RulesConceptsAndRelations;
